PARTIE3: CH???5&6???? i dont know, i just messed up, and repared it after BUT THE EXEMPLE OF THE "ERROR" JUST BEFORE THE QUIZ, HE DID NOT FINISH, I HAD TO ADD AN if CONDITION TO DON'T CRASH THE SERVER, BUT STILL THE RESPONSE DOES NOT SCREEN/APPEAR CLEAN

+++ controllers/stuff.js
@@ -64,19 +64,38 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({_id: req.params.id}).then(
-    () => {
-      res.status(200).json({
-        message: 'Deleted!'
-      });
-    }
-  ).catch(
-    (error) => {
-      res.status(400).json({
-        error: error
-      });
-    }
-  );
+    Thing.findOne({ _id: req.params.id }).then(
+        (thing) => {
+            if (!thing) {
+                console.log("error, no 'thing' have been returned");
+                res.status(404).json({
+                    error: new Error('No such Thing!')
+                });
+            }
+            else if (thing.userId !== req.auth.userId) {
+                console.log("error, userId doesn't match\nthing.userId: "+thing.userId+" --- req.auth.userId: "+req.auth.userId);
+                res.status(400).json({
+                    error: new Error('Unauthorized request!')
+                });
+            }
+            else{
+                console.log("oh heyn its good, it did match!\nthing.userId: "+thing.userId+" --- req.auth.userId: "+req.auth.userId);
+                Thing.deleteOne({ _id: req.params.id }).then(
+                    () => {
+                        res.status(200).json({
+                            message: 'Deleted!'
+                        });
+                    }
+                ).catch(
+                    (error) => {
+                        res.status(400).json({
+                            error: error
+                        });
+                    }
+                );
+            }
+        }
+    )
 };
 
 exports.getAllStuff = (req, res, next) => {
